Expose monitorRegion Move extension to MKCoordinateRegion to WebSocketManager
