+++ Sources/Core/Web Socket/WebSocketManager.swift
@@ -13,6 +13,7 @@ import os
 import UIKit
 import Starscream
 import Signals
+import MapKit
 
 /*
  Important points:
@@ -289,6 +290,15 @@ public class WebSocketManager {
         self.write(commandPackage)
 
     }
+    
+    /// Begins region monitoring on the specified `MKCoordinateRegion`.
+    ///
+    /// Only a single region can be monitored at a time. Calling this method will update the monitored region.
+    ///
+    /// - Parameter region: The `MKCoordinateRegion` to monitor.
+    public func monitorRegion(_ region: MKCoordinateRegion) {
+        self.writeRegionCommand(region.toDictionary())
+    }
 
     // MARK: Debugging
 
@@ -444,6 +454,25 @@ extension WebSocketManager: WebSocketDelegate {
             return
         }
 
+    }
+
+}
+
+extension MKCoordinateRegion {
+
+    /// Returns a dictionary in data pool format.
+    func toDictionary() -> [String: Any] {
+        let payload: [String: [String: Any]] = [
+            "top_left": [
+                "lat": self.topLeft.latitude,
+                "lon": self.topLeft.longitude
+            ],
+            "bottom_right": [
+                "lat": self.bottomRight.latitude,
+                "lon": self.bottomRight.longitude
+            ]
+        ]
+        return payload
     }
 
 }
